[REF] connector_sage: move test adapter class
